requirements.txt
atomicwrites==1.4.0
attrs==21.4.0
autopep8==1.6.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
cycler==0.10.0
flake8==4.0.1
flake8-broken-line==0.4.0
flake8-isort==4.1.1
flake8-plugin-utils==1.3.2
flake8-polyfill==1.0.2
flake8-return==1.1.3
idna==3.3
importlib-metadata==4.2.0
iniconfig==1.1.1
isort==5.10.1
kiwisolver==1.3.1
matplotlib==3.3.3
mccabe==0.6.1
mypy==0.942
mypy-extensions==0.4.3
numpy==1.19.5
packaging==21.3
pep8-naming==0.12.1
Pillow==8.1.0
pluggy==1.0.0
prettytable==2.0.0
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pygame==1.9.6
pyglet==1.5.8
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
pytest==7.1.1
python-dateutil==2.8.1
python-dotenv==0.20.0
qt5-applications==5.15.2.2.2
requests==2.27.1
six==1.15.0
soupsieve==2.3.2.post1
testfixtures==6.18.5
toml==0.10.2
tomli==2.0.1
typed-ast==1.5.3
typing_extensions==4.2.0
urllib3==1.26.9
wcwidth==0.2.5
zipp==3.8.0
